infra/cdk/lib/lambdas/prelander_image_gen/pyproject.toml
[project]
name = "deepcopy-prelander-image-gen"
version = "0.1.0"
description = "Lambda prelander image generation processor"
requires-python = ">=3.12"
dependencies = [
  "requests>=2.31.0",
  "cloudflare>=4.0.0",
  "pillow>=10.0.0",
  "google-genai>=1.0.0",
  "boto3>=1.26.0",
]

[tool.setuptools]
py-modules = ["handler", "llm_usage"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

